+++ tcex/api/tc/v2/batch/batch_writer.py
@@ -20,6 +20,7 @@ from requests import Session  # TYPE-CHECKING
 
 # first-party
 from tcex.api.tc.util.threat_intel_util import ThreatIntelUtil
+from tcex.api.tc.v2.batch.association import Association
 from tcex.api.tc.v2.batch.group import (
     Adversary,
     AttackPattern,
@@ -136,6 +137,7 @@ class BatchWriter:
         self._groups_shelf = None
         self._indicators = None
         self._indicators_shelf = None
+        self._associations = None
 
         # build custom indicator classes
         self._gen_indicator_class()
@@ -407,6 +409,13 @@ class BatchWriter:
                 indicator_data['flag2'] = whois_active
 
         return self._indicator(indicator_data, kwargs.get('store', True))
+
+    def add_association(self, association: Association | dict) -> Association:
+        """Add an association to Batch Job."""
+        if isinstance(association, dict):
+            association = Association(**association)
+        self.associations.add(association)
+        return association
 
     def address(self, ip: str, **kwargs) -> Address | dict:
         """Add Address data to Batch.
@@ -899,6 +908,13 @@ class BatchWriter:
             # new shelf file
             self._group_shelf_fqfn = self.inputs.model.tc_temp_path / f'groups-{uuid.uuid4()!s}'
         return self._group_shelf_fqfn
+
+    @property
+    def associations(self) -> dict:
+        """Return dictionary of all Associations data."""
+        if self._associations is None:
+            self._associations = set()
+        return self._associations
 
     @property
     def groups(self) -> dict: